comfyui-tenant-service/pyproject.toml
[tool.poetry]
name = "comfyui-tenant-service"
version = "0.1.0"
description = "Multi-tenant microservice for ComfyUI Runninghub"
authors = ["JOTO Team <dev@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
httpx = "^0.27.2"
python-multipart = "^0.0.9"
pydantic-settings = "^2.5.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
sqlite3 = "^0.4.6"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.9"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
